moved total costs to it's own section
